dualnback/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
channels==4.0.0
channels-redis==4.2.1
daphne==4.1.2
django-cors-headers==4.7.0
django-filter==25.1
redis==6.0.0
whitenoise==6.6.0